+++ src/app/shared/services/elapsed-time.service.ts
@@ -5,11 +5,11 @@ import { Store } from '@ngrx/store';
 import { combineLatest, interval, Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 
-import sharedSelectors, { SharedState } from '../../../shared/reducers/root.reducer';
+import sharedSelectors, { SharedState } from '../reducers/root.reducer';
 
-import { TimerInfo } from '../../../shared/models';
+import { TimerInfo } from '../models';
 
-import { formatElapsedTime } from '../../../shared/utils/date.utils';
+import { formatElapsedTime } from '../utils/date.utils';
 
 @Injectable()
 export class ElapsedTimeService {
